# Conflicts:
#	CHANGELOG.md
#	core-bundle/src/Resources/contao/config/constants.php
#	core-bundle/src/Resources/contao/models/PageModel.php
#	core-bundle/src/Resources/contao/templates/backend/be_tinyMCE.html5

+++ news-bundle/src/Resources/contao/modules/ModuleNewsArchive.php
@@ -169,7 +169,7 @@ class ModuleNewsArchive extends ModuleNews
 
 				// Get the current page
 				$id = 'page_a' . $this->id;
-				$page = Input::get($id) ?? 1;
+				$page = (int) (Input::get($id) ?? 1);
 
 				// Do not index or cache the page if the page number is outside the range
 				if ($page < 1 || $page > max(ceil($total/$this->perPage), 1))
